fix: spawn user in scene base coords
