+++ src/lib/Decentraland.ts
@@ -10,9 +10,9 @@ import { ContentService } from './content/ContentService'
 import { filterAndFillEmpty } from '../utils/land'
 import { Coords, getObject } from '../utils/coordinateHelpers'
 import { ErrorType, fail } from '../utils/errors'
-import { DCLInfo, getConfig } from '../config'
+import { DCLInfo, getConfig, SceneMetadata } from '../config'
 import { warning, debug } from '../utils/logging'
-import { Project, BoilerplateType, IFile, SceneMetadata } from './Project'
+import { Project, BoilerplateType, IFile } from './Project'
 import { Ethereum, LANDData } from './Ethereum'
 import { LinkerAPI, LinkerResponse } from './LinkerAPI'
 import { Preview } from './Preview'
